refactor: restructure venus-subgraph to match others
