+++ subgraphs/venus/src/constants/addresses.ts
@@ -1,0 +1,2 @@
+export const vBNBAddress = '0xa07c5b74c9b40447a954e1466938b865b6bbea36';
+export const vUSDCAddress = '0xeca88125a5adbe82614ffc12d0db554e2e2867c8';
